Move relationship key filtering out into util
